panel/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
